Convert Google Slides and notes to video with ari

Convert Google Slides and notes to video with ari

gs_to_ari(path, script = NULL, ..., verbose = TRUE)

pptx_to_pdf(path, verbose = TRUE)

pptx_to_pngs(path, verbose = TRUE, dpi = 600)

pptx_to_ari(path, script = NULL, ..., verbose = TRUE)

pdf_to_ari(path, script = NULL, dpi = 300, ..., verbose = TRUE)

pdf_to_pngs(path, verbose = TRUE, dpi = 600)

images_to_ari(path, script = NULL, dpi = 300, ..., verbose = TRUE)

to_ari(path, script = NULL, ..., verbose = TRUE)

Arguments

path
Identifier of google slides presentation, or PPTX filename

script
passed to make_ari_document()

...
Arguments passed to make_ari_document

verbose
print diagnostic messages

dpi
resolution (dots per inch) to render images

Value

The output from make_ari_document

Examples

# \donttest{
id = "1Opt6lv7rRi7Kzb9bI0u3SWX1pSz1k7botaphTuFYgNs"
res = gs_to_ari(id, verbose = 2, open = FALSE)

#> Downloading PPTX

#> Warning: This presentation may not be available, did you turn link sharing on?

#> pptx is at: /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd138bc68e.pptx

#> Getting Notes from PPTX

#> script is at: /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd6ed8472f.txt

#> Downloading PDF

#> Warning: This presentation may not be available, did you turn link sharing on?

#> PDF is at: /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd39299783.pdf

#> Converting PDF to PNGs

#> Converting page 1 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd599a36c6.png... done!
#> Converting page 2 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd1b822f1f.png... done!
#> Converting page 3 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd7fe3aa54.png... done!
#> Converting page 4 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd3bc3b264.png... done!
#> Converting page 5 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd2cf4e1e3.png... done!
#> Converting page 6 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd11a1524.png... done!
#> Converting page 7 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd5761ef0c.png... done!

#> output_file is: /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd5e9529a5.md

#> Making output_file directories

#> output_dir is at: /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpHV0paS

#> files_dir is at: /private/var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T/RtmpHV0paS/file59dd5e9529a5_files

#> Creating YAML header

if (interactive()) {
file.edit(res$output_file)
}
res2 = to_ari(id, open = FALSE)

#> Downloading PPTX

#> Warning: This presentation may not be available, did you turn link sharing on?

#> Getting Notes from PPTX

#> Downloading PDF

#> Warning: This presentation may not be available, did you turn link sharing on?

#> Converting PDF to PNGs

#> Converting page 1 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd277d139d.png... done!
#> Converting page 2 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd492bcac.png... done!
#> Converting page 3 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd3da0c28c.png... done!
#> Converting page 4 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd54494f0.png... done!
#> Converting page 5 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd598e1b43.png... done!
#> Converting page 6 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd9ff9a4.png... done!
#> Converting page 7 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd6be7e4e.png... done!

#> Making output_file directories

# }

bad_id = "1AJxokPkGVzNZMXjyF59mNe3EiQRGxylbBfYGGoQMQls"
testthat::expect_error(res = gs_to_ari(bad_id))

ex_file = system.file("extdata", "example.pptx", package = "ariExtra")
have_soffice = try(docxtractr:::lo_assert())
if (!inherits(have_soffice, "try-error")) {
pngs = pptx_to_pngs(ex_file)
res = pptx_to_ari(ex_file, open = FALSE)
if (interactive()) {
file.edit(res$output_file)
}
res2 = to_ari(ex_file, open = FALSE)
}

#> Converting PPTX to PDF

#> Converting PDF to PNGs

#> Converting page 1 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd56a3c0a.png... done!
#> Converting page 2 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd18bb74d.png... done!

#> Getting Notes from PPTX

#> Converting PPTX to PDF

#> Converting PDF to PNGs

#> Converting page 1 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd338af32e.png... done!
#> Converting page 2 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd67636b71.png... done!

#> Making output_file directories

#> Getting Notes from PPTX

#> Converting PPTX to PDF

#> Converting PDF to PNGs

#> Converting page 1 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd66737328.png... done!
#> Converting page 2 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd258d7ba4.png... done!

#> Making output_file directories

ex_file = system.file("extdata", "example.pdf", package = "ariExtra")
res = pdf_to_ari(ex_file, script = c("hey", "ho"), open = FALSE)

#> Converting PDF to PNGs

#> Converting page 1 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd83d16ae.png... done!
#> Converting page 2 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd629bfdbb.png... done!

#> Making output_file directories

if (interactive()) {
file.edit(res$output_file)
}
res2 = to_ari(ex_file,  script = c("hey", "ho"), open = FALSE)

#> Converting PDF to PNGs

#> Converting page 1 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd3a0c4f7.png... done!
#> Converting page 2 to /var/folders/1s/wrtqcpxn685_zk570bnx9_rr0000gr/T//RtmpHV0paS/file59dd2fe335fd.png... done!

#> Making output_file directories